# Conflicts:
#	rllib/algorithms/algorithm_config.py
#	rllib/algorithms/tests/test_worker_failures.py
#	rllib/env/remote_base_env.py
#	rllib/env/wrappers/model_vector_env.py
#	rllib/evaluation/collectors/agent_collector.py
#	rllib/evaluation/env_runner_v2.py
#	rllib/evaluation/rollout_worker.py
#	rllib/evaluation/sampler.py
#	rllib/evaluation/tests/test_rollout_worker.py
#	rllib/tests/test_pettingzoo_env.py
#	rllib/tuned_examples/a2c/atari-a2c.yaml
#	rllib/tuned_examples/ddpg/pendulum-ddpg.yaml
#	rllib/tuned_examples/sac/cartpole-continuous-pybullet-sac.yaml

+++ rllib/algorithms/dt/dt_torch_policy.py
@@ -536,7 +536,9 @@ class DTTorchPolicy(LearningRateSchedule, TorchPolicyV2):
             Scalar cross entropy loss.
         """
         losses = F.cross_entropy(
-            preds.reshape(-1, preds.shape[-1]), targets.reshape(-1), reduction="none"
+            preds.reshape(-1, preds.shape[-1]),
+            targets.reshape(-1).long(),
+            reduction="none",
         )
         losses = losses * masks.reshape(-1)
         return losses.mean()
